Move test outside of application code and rename main source directory
Using the layout suggested here
https://docs.pytest.org/en/stable/explanation/goodpractices.html#tests-outside-application-code
