+++ Containerfile
@@ -46,11 +46,11 @@ RUN microdnf install -y --nodocs \
     "python${PYTHON_VERSION}" \
     && rm -rf /var/cache/yum/*
 
-COPY app/ /srv/rhel_roadmap/app/
+COPY roadmap/ /srv/rhel_roadmap/roadmap/
 
 RUN useradd --system --create-home --home-dir /srv/rhel_roadmap roady
 
 USER roady
 WORKDIR /srv/rhel_roadmap
 
-CMD ["fastapi", "run", "app/main.py", "--proxy-headers", "--port", "80"]
+CMD ["fastapi", "run", "roadmap/main.py", "--proxy-headers", "--port", "80"]
