+++ roadmap/v1/lifecycle/systems.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, Path
 
-from app.data.systems import OS_DATA_MOCKED
+from roadmap.data.systems import OS_DATA_MOCKED
 
 v1_router = APIRouter(
     prefix="/systems",
