+++ tests/conftest.py
@@ -1,14 +1,9 @@
 import pytest
 from fastapi.testclient import TestClient
 
-from app.main import app
+from roadmap.main import app
 
 
 @pytest.fixture
 def client():
     return TestClient(app)
-
-
-@pytest.fixture(scope="session")
-def api_prefix():
-    return "/api/digital-roadmap/v1"
