+++ tests/test_main.py
@@ -1,5 +1,5 @@
-def test_ping(client, api_prefix):
-    response = client.get(f"{api_prefix}/ping")
+def test_ping(client):
+    response = client.get("/api/digital-roadmap/v1/ping")
 
     assert response.status_code == 200
     assert response.json() == {"status": "pong"}
