+++ tests/v1/conftest.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.fixture(scope="session")
+def api_prefix():
+    return "/api/digital-roadmap/v1"
