+++ tests/v1/lifecycle/test_app_streams.py
